Move tests from base to base_circuit
