+++ src/qibo/tests/test_base.py
@@ -29,85 +29,6 @@ def test_importing_qibo_modules():
     assert c.ngates == 1
 
 
-def test_gate_types():
-    """Check ``BaseCircuit.gate_types`` property."""
-    import collections
-    c = Circuit(3)
-    c.add(H(0))
-    c.add(H(1))
-    c.add(X(2))
-    c.add(CNOT(0, 2))
-    c.add(CNOT(1, 2))
-    c.add(TOFFOLI(0, 1, 2))
-    target_counter = collections.Counter({"h": 2, "x": 1, "cx": 2, "ccx": 1})
-    assert target_counter == c.gate_types
-
-
-def test_gates_of_type():
-    """Check ``BaseCircuit.gates_of_type`` method."""
-    c = Circuit(3)
-    c.add(H(0))
-    c.add(H(1))
-    c.add(CNOT(0, 2))
-    c.add(X(1))
-    c.add(CNOT(1, 2))
-    c.add(TOFFOLI(0, 1, 2))
-    c.add(H(2))
-    h_gates = c.gates_of_type(H)
-    cx_gates = c.gates_of_type("cx")
-    assert h_gates == [(0, c.queue[0]), (1, c.queue[1]), (6, c.queue[6])]
-    assert cx_gates == [(2, c.queue[2]), (4, c.queue[4])]
-    with pytest.raises(TypeError):
-        c.gates_of_type(5)
-
-
-def test_summary():
-    """Check ``BaseCircuit.summary()`` method."""
-    c = Circuit(3)
-    c.add(H(0))
-    c.add(H(1))
-    c.add(CNOT(0, 2))
-    c.add(CNOT(1, 2))
-    c.add(TOFFOLI(0, 1, 2))
-    c.add(H(2))
-    target_summary = "\n".join(["Circuit depth = 5",
-                                "Total number of gates = 6",
-                                "Number of qubits = 3",
-                                "Most common gates:",
-                                "h: 3", "cx: 2", "ccx: 1"])
-    assert c.summary() == target_summary
-
-
-@pytest.mark.parametrize("deep", [False, True])
-def test_circuit_copy(deep):
-    """Check that ``circuit.copy()`` copies gates properly."""
-    c1 = Circuit(2)
-    c1.add([H(0), H(1), CNOT(0, 1)])
-    c2 = c1.copy(deep)
-    assert c2.depth == c1.depth
-    assert c2.ngates == c1.ngates
-    assert c2.nqubits == c1.nqubits
-    for g1, g2 in zip(c1.queue, c2.queue):
-        if deep:
-            assert g1.__class__ == g2.__class__
-            assert g1.target_qubits == g2.target_qubits
-            assert g1.control_qubits == g2.control_qubits
-        else:
-            assert g1 is g2
-
-
-def test_circuit_copy_with_measurements():
-    """Check that ``circuit.copy()`` copies measurements properly."""
-    c1 = Circuit(4)
-    c1.add([H(0), H(3), CNOT(0, 2)])
-    c1.add(M(0, 1, register_name="a"))
-    c1.add(M(3, register_name="b"))
-    c2 = c1.copy()
-
-    assert c2.measurement_gate is c1.measurement_gate
-    assert c2.measurement_tuples == {"a": (0, 1), "b": (3,)}
-
-
 def test_base_gate_errors():
     """Check errors in ``base.gates.Gate`` for coverage."""
     gate = H(0)
